StandaloneMmPkg/Core: dump HOB list in MMRAM
This commit is to dump finaled HOB list in MMRAM.

In previous commit, we verify the HOB list and initialize a
new HOB list in MMRAM before we call the library constructor.
Since we might migrate some memory allocation HOB into MMRAM,
it's cleared to dump information in finaled HOB list.

Signed-off-by: Dun Tan <dun.tan@intel.com>