MdeModulePkg/BootManagerUiLib: hide UiApp from boot entries
After a57f67938435 ("OvmfPkg/PlatformBootManagerLib: Register UiApp as
an optional boot option") UiApp can be found among the boot options
without the LOAD_OPTION_HIDDEN flag set. This means that it can appear
in the list of available entries. This can be confusing.

Starting from the UiApp guid, add a check on the Device Path on all the
available entries, if it matches, skip the entry.

Signed-off-by: Luigi Leonardi <leonardi@redhat.com>