BaseTools: Add support for GCC preprocessor line directives
This change adds support for GCC-style preprocessor line
directives as documented in:
https://gcc.gnu.org/onlinedocs/cpp/Preprocessor-Output.html

On Windows systems, one can use line-markers to see  which .vfr
file was used to generate a *.i file in the Build directory.
This is useful for debugging VFR compilation failures.

With this change, the VfrCompiler will not generate compilation
errors if the *.i file contains GCC-style line-markers. Without
this change, one must disable the pre-processor's emission of
line-markers, removing the debug aid they provide.

Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>