MdeModulePkg: Fix Image Memory Protection Applying
Commit 12d4204ca42a5e512c24b60497069ebf5e951b7d updated the
image memory protection code to set the protection
attributes through the GCD instead of directly to the page
table. However, this code had an implicit assumption that
each base address passed to it was the beginning of a GCD
descriptor. On the virtual platforms tested, this was the case.
However, on a physical platform, a scenario was encountered
where the base address was not the beginning of a GCD
descriptor, thus causing memory attributes to be applied
incorrectly.

This assumption does not need to be made and this patch
updates the code to handle the case where the base address
is not the beginning of a GCD descriptor.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>