DynamicTablesPkg: Add CmObjHelperLib
Some generators share some logic in how they handle CmObj.
In an effort to factorize this code, create a CmObjHelperLib,
aiming to host the common CmObj manipulation logic.

Start by adding a CheckAcpiTablePresent() function, allowing
to check if an ACPI table is in the list of the tables to
install.

Signed-off-by: Pierre Gondois <pierre.gondois@arm.com>