OvmfPkg/CpuHotplugSmm: set MSR_IA32_FEATURE_CONTROL in first SMI handler
Introduce new fields to FIRST_SMI_HANDLER_CONTEXT:

- a UINT8 (effectively: boolean) value that controls whether the first SMI
  handler of the hot-added CPU is supposed to set
  MSR_IA32_FEATURE_CONTROL,

- and for when that is the case, two UINT32 values, for passing the
  desired EDX:EAX (64-bit) MSR value.

MSR_IA32_FEATURE_CONTROL might as well be settable in the post-SMM pen
("PostSmmPen.nasm"); however, I find the first SMI handler better for this
purpose:

- we already have a nice context structure we can extend,

- in SMM, we have tighter control than after SMM (see the comments near
  the top of "PostSmmPen.nasm").

MSR_IA32_FEATURE_CONTROL is orthogonal to the SMRAM State Save Map, and
WRMSR is permitted in the real mode-like initial environment of SMM, so
functionally speaking, this should be fine.

For now, the feature is disabled.

Signed-off-by: Laszlo Ersek <laszlo.ersek@posteo.net>